Move MongoConnectedTestBase into the library
This is a useful base class for functional tests that need to verify stuff against a MongoDB, so we might as well make it available.
