+++ test/functional/MongoManager_EnsureIndexes_Tests.cs
@@ -4,6 +4,7 @@ using System.Reflection;
 using MongoDB.Bson;
 using MongoDB.Driver;
 using RapidCore.Mongo.Internal;
+using RapidCore.Mongo.Testing;
 using Xunit;
 
 namespace RapidCore.Mongo.FunctionalTests
